Fix: signed integer overflow when mixing in server seed value.
sysopt.serverseed is a plain int filled across the full positive
range [1, INT_MAX] (the install-time generator uses v % 2147483647
and read_serverseed_file() accepts any val up to INT_MAX), so signed
arithmetic that adds to it can overflow. On a build compiled with
-fsanitize=undefined -fno-sanitize-recover=all this aborts the game
outright; a player hit it at new-game start, where any seed in
(INT_MAX - 8000000, INT_MAX] overflowed. The only real consumer,
coord_hash(), already reads the value as unsigned int, so the
overflow served no functional purpose.

* src/allmain.c: newgame() did sysopt.serverseed += rn2(8000000),
  overflowing int for seeds near INT_MAX and aborting under UBSan at
  new-game start. Compute the sum as unsigned and mask with
  0x7fffffff so the result stays a non-negative int.

* src/shknam.c: the shopkeeper-name seed (int)((long)ubirthday/257L)
  + sysopt.serverseed had the same latent overflow. Promote the
  addition to unsigned and cast back; nseed is used only via % 13
  and % 5 with an existing negative-result guard, so the defined
  wraparound is harmless and the non-overflowing value is unchanged.

+++ src/shknam.c
@@ -514,7 +514,8 @@ const char *const *nlp;
            and restore support which would be necessary for randomization;
            try not to make too many assumptions about time_t's internals;
            use ledger_no rather than depth to keep mine town distinct. */
-        int nseed = (int) ((long) ubirthday / 257L) + sysopt.serverseed;
+        int nseed = (int) ((unsigned) ((long) ubirthday / 257L)
+                           + (unsigned) sysopt.serverseed);
 
         name_wanted = ledger_no(&u.uz) + (nseed % 13) - (nseed % 5);
         if (name_wanted < 0)
